+++ edge-raw/src/lib.rs
@@ -66,8 +66,7 @@ impl defmt::Format for Error {
     }
 }
 
-#[cfg(feature = "std")]
-impl std::error::Error for Error {}
+impl core::error::Error for Error {}
 
 /// Decodes an IP packet and its UDP payload
 #[allow(clippy::type_complexity)]