+++ src/lib.rs
@@ -5,8 +5,6 @@ pub use edge_captive as captive;
 pub use edge_dhcp as dhcp;
 pub use edge_http as http;
 pub use edge_mdns as mdns;
-#[cfg(feature = "std")]
-pub use edge_mqtt as mqtt;
 #[cfg(feature = "io")]
 pub use edge_nal as nal;
 #[cfg(feature = "embassy")]