+++ edge-mdns/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2026-01-01
+* Update the `edge-nal` and `embassy-time` dependencies
+* Update the `domain` dependency to 0.11
+* Addressed #60
+
 ## [0.6.1] - 2025-08-14
 * Fix HostAnswersMdnsHandler responding to any A/AAAA queries even without match