+++ edge-nal-std/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2026-01-01
+* Update the `edge-nal` and `embedded-io-async` dependencies
+* Use `async-io-mini` from `crates.io`
+
 ## [0.5.1] - 2025-11-23
 * Document the API
 * Make `Stack` `Copy`